pyproject.toml
[project]
name = "matvey-3000"
version = "0.15.0"
description = "OpenAI Telegram bot responder with multi-provider LLM support"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.6.0",
    "anthropic>=0.16.0",
    "hiredis>=2.3.2",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "openai>=1.30.0",
    "Pillow>=10.3.0",
    "redis>=5.0.4",
    "tiktoken>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.uv]
package = false
